Move mobile landing to unauthenticated
