+++ www/src/App.tsx
@@ -10,10 +10,10 @@ import { Header, Section } from './components/Header';
 import { useAppDispatch, useAppSelector } from './redux/hooks';
 import adminStore from './redux/substores/admin/adminStore';
 import checkUserRegistration from './redux/thunks/checkUserRegistration';
-import MobileLanding from './routes/MobileLanding';
 import StudentApp from './routes/Student';
 import Register from './routes/student/Register';
 import UnauthenticatedApp from './routes/Unauthenticated';
+import MobileLanding from './routes/unauthenticated/MobileLanding';
 import VolunteerApp from './routes/Volunteer';
 import theme from './styles/theme';
 import { COMMUNITY, COMMUNITY_ROUTE, COMPE_PLUS, MOCK_INTERVIEW, MOCK_INTERVIEW_ROUTE, REGISTER_ROUTE, RESUME_REVIEW, RESUME_REVIEW_ROUTE } from './util/constants';
